requirements.txt
asgiref==3.2.3
Django==3.0.1
django-cors-headers==3.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2019.3
reverse==0.1.0
sqlparse==0.3.0
django-rest-registration